% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsummary.R
\name{nsummary}
\alias{nsummary}
\title{Creates histogram, boxplot, and numeric summary}
\usage{
nsummary(data)
}
\arguments{
\item{data}{numeric variable}
}
\description{
Creates histogram, boxplot, and numeric summary
}
